src/pyproject.toml
[tool.poetry]
name = "foobarqix"
version = "0.1.0"
description = ""
authors = ["r2r-dev <dev@r2r.sh>"]

[tool.poetry.dependencies]
python = "^3.7"
fastapi = "^0.63.0"
uvicorn = "^0.13.2"
gunicorn = "^20.0.4"
uvloop = "^0.14.0"
httptools = "^0.1.1"
click = "^7.1.2"

[tool.poetry.dev-dependencies]
autoflake = "^1.4"
black = "^20.8b1"
flake8 = "^3.8.4"
flake8-fixme = "^1.1"
isort = "^4.3.21"
pytest = "^6.2.1"
pytest-cov = "^2.10.1"
pytest-parametrized = "^1.3"
requests = "^2.25.1"
wemake-python-styleguide = "^0.14.1"

[tool.isort]
profile = "black"
src_paths = ["app", "tests"]
combine_as_imports = true

[tool.pytest.ini_options]
testpaths = "tests"
filterwarnings = "error"
addopts = '''
  --strict-markers
  --tb=short
  --cov=app
  --cov=tests
  --cov-branch
  --cov-report=term-missing
  --cov-report=html
  --cov-report=xml
  --no-cov-on-fail
  --cov-fail-under=100
'''

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
